finished shadow pass

+++ Framework/ECS/Components/Scene/AspectRatioComponent.cs
@@ -9,8 +9,8 @@ namespace Framework.ECS.Components.Scene
     [DebuggerDisplay("Ratio: {Ratio}, Width: {Width}, Height: {Height}")]
     public struct AspectRatioComponent
     {
-        public float Width { get; set; }
-        public float Height { get; set; }
-        public float Ratio => Width / Height;
+        public int Width { get; set; }
+        public int Height { get; set; }
+        public float Ratio => Width / (float)Height;
     }
 }
